+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * updated dependencies
 * updated referrer blacklist
 * map Android referrers to app name and icon from the Google Play Store
+* added bounce rate per referrer
 
 ## 1.10.7